+++ packages/tools/geometry/src/index.js
@@ -1,4 +1,5 @@
 export * from './2DNetwork.js';
 export * from './2DPoint.js';
 export * from './2DVector.js';
-export * from './3D.js';
+export * from './3DPoint.js';
+export * from './angle.js';